Create test_LeapYearFileTestDriven.py
Created test file for a program that decides wether a given year will be a leap year. First test checks to see if the year is both divisible by 100 and 400, if it is, it is not a leap year. With the given input of 2400, the function should return "False," to indicate that the year is NOT a leap year.